requirements.txt
click~=8.1.7
certifi~=2024.8.30
mongoengine~=0.29.0
pymongo~=4.10.1
metroapi~=0.1
uvicorn~=0.30.6
aioredis~=2.0.1
pydantic~=2.10.4
requests~=2.32.3
httpx~=0.27.2
jinja2~=3.1.4
inflect~=7.4.0
bcrypt~=4.2.0
cryptography~=43.0.1
python-dotenv~=1.0.1
fastapi~=0.114.2
starlette~=0.38.5
setuptools~=65.5.1
bson~=0.5.10
PyJWT~=2.10.1
redis~=5.1.0
twilio~=9.4.1
vonage~=4.1.2